+++ website/src/js/components/BlocksPage/GFXElems/GFXFlourish.jsx
